file moving cleanup
